forge build works

+++ contracts/src/helpers/Math.sol
@@ -3,15 +3,15 @@ pragma solidity ^0.8.13;
 
 // a library for performing various math operations
 library Math {
-    function min(uint x, uint y) internal pure returns (uint z) {
+    function min(uint256 x, uint256 y) internal pure returns (uint256 z) {
         z = x < y ? x : y;
     }
 
     // babylonian method (https://en.wikipedia.org/wiki/Methods_of_computing_square_roots#Babylonian_method)
-    function sqrt(uint y) internal pure returns (uint z) {
+    function sqrt(uint256 y) internal pure returns (uint256 z) {
         if (y > 3) {
             z = y;
-            uint x = y / 2 + 1;
+            uint256 x = y / 2 + 1;
             while (x < z) {
                 z = x;
                 x = (y / x + x) / 2;
@@ -21,14 +21,19 @@ library Math {
         }
     }
 
-    function mulDiv (uint x, uint y, uint z)
-      public pure returns (uint) {
-        return (x * y)/ z;
+    function mulDiv(
+        uint256 x,
+        uint256 y,
+        uint256 z
+    ) public pure returns (uint256) {
+        return (x * y) / z;
     }
 
-
-    function calculateInterest (uint principal, uint interest, uint period )
-      public pure returns (uint) {
+    function calculateInterest(
+        uint256 principal,
+        uint256 interest,
+        uint256 period
+    ) public pure returns (uint256) {
         return mulDiv(principal, mulDiv(interest, period, 365), 10**18);
     }
 }
